fix(hermes): address PR #973 reviewer feedback for recommend
- [P1] Split runner functions out of cli_hermes.py into new
  cli_hermes_runners.py. cli_hermes.py now 357 lines (was 901);
  cli_hermes_runners.py 619 lines. Both under the 800-LOC guard.
  _print_inventory moved alongside its sole caller. Public surface
  unchanged: cli_hermes.py still owns register_hermes_command and
  the typer subcommand wiring; runners is the pure-Python body the
  subcommands call.
- [P2] Read-only invariant: `autoctx hermes recommend --output`
  paths inside the resolved Hermes home are rejected at the
  boundary. New _is_inside() helper uses Path.resolve() +
  is_relative_to() so symlinks and nested directories cannot
  bypass. Without this, --output ~/.hermes/recs.jsonl would create
  a file under the home, breaking the AC-709 read-only contract.

Adds 2 regression tests (output directly under home; output in
nested subdir). 191 hermes tests pass; lint + mypy clean; module
size limits test passes.
